Update sanoid_healthcheck
These are my changes and fixes from the initial commit.

There was an error on line 28 in the original, 39 in this version, that was solved by adding quotes. 

Updated the logging to truncate, not append

Cleaned up the output in the log file to suit my preferences, and be more visually useful to me.

I added some new variables to make the script a little easier for me to customize for each server I'll run it from.